Add Libreoffice to default substitutions
Window Title Applet displays LibreOffice's name as "soffice.bin", which is ugly. "LibreOffice" is much better. It is quite a common application (probably more so than Telegram Desktop:-)). Actually, Plasma displays it correctly with the program, so it shows "LibreOffice - Writer", not just LibreOffice. (well, there should be an emdash, but that is somebody else's bug).