Change plasmapkg2 to kpackagetool5 in readme
On some machines (e.g. my Manjaro KDE one), `plasmapkg2` is no longer available. On others, it's simply an alias for `kpackagetool5`. For example, on Fedora, `plasmapkg -v` outputs `kpackagetool5 2.0`. Usage appears to be exactly the same, so there shouldn't be any harm in replacing it - especially because the same line in the README mentions `Plasma 5`, which always comes with `kpackagetool5`.
BTW, thanks for creating and maintaining the app, starred :)